Move io readers into seperate module
Does not touch the public interface, only internal `_impl` variants
which are simply referenced from their new location.
