+++ src/io/mod.rs
@@ -1,0 +1,2 @@
+//! Input and output of images.
+pub(crate) mod free_functions;
