+++ src/lib.rs
@@ -61,7 +61,8 @@ pub use flat::{FlatSamples};
 pub use traits::Primitive;
 
 // Opening and loading images
-pub use dynimage::{guess_format, load, load_from_memory, load_from_memory_with_format, open,
+pub use io::free_functions::{guess_format, load};
+pub use dynimage::{load_from_memory, load_from_memory_with_format, open,
                    save_buffer, save_buffer_with_format, image_dimensions};
 
 pub use dynimage::DynamicImage::{self, ImageLuma8, ImageLumaA8, ImageRgb8, ImageRgba8, ImageBgr8, ImageBgra8};
@@ -73,6 +74,9 @@ pub mod math;
 
 // Image processing functions
 pub mod imageops;
+
+// Io bindings
+pub mod io;
 
 // Buffer representations for ffi.
 pub mod flat;
